move GetInitType to daemon package
moved CheckConfig from config pkg to cmd to eliminate cyclical
dependency
