+++ daemon/common_freebsd.go
@@ -195,3 +195,8 @@ func removeFreebsdDaemon() error {
 	}
 	return nil
 }
+
+// GetInitType - returns init type (not applicable for freebsd)
+func GetInitType() config.InitType {
+	return config.UnKnown
+}
