+++ daemon/common_windows.go
@@ -139,3 +139,8 @@ func runWinSWCMD(command string) error {
 	}
 	return err
 }
+
+// GetInitType - returns the init type (not applicable for windows)
+func GetInitType() config.InitType {
+	return config.UnKnown
+}
